requirements.txt
flask==2.3.3
pillow==10.0.0
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.24.3
transformers==4.31.0
torch==2.0.1
sentencepiece==0.1.99
protobuf==3.20.3
pdf2image==1.16.3
python-multipart==0.0.6
waitress==2.1.2